Cache geocoding results in manager object cache
The move of the cache makes restarts without losing
cached geocoding results possible; if a disk cache is used.
Previously the cache was always in-memory only.

Cached results expire after one month now.

The new 'gmaps-cache-fuzz' option can be used in the config
to flatten re-lookup spikes that could happen if a lot of
requests were made at the same day, i.e. during an event.
It randomly increases cache expiration to up to the configured
value in days (defaults to 0).